requirements.txt

flask
openpyxl
gunicorn
sqlalchemy
psycopg2-binary
python-dotenv

